Bouton confirmation delete

+++ view/private/admin.homepage.html.php
@@ -39,7 +39,7 @@
                             <td><?=$item['latitude']?></td>
                             <td><?=$item['longitude']?></td>
                             <td><a href="?update=<?=$item['idourdatas']?>"><img src="img/update.png" alt="icon update" width="32" height="32"></a></td>
-                            <td><a href="?delete=<?=$item['idourdatas']?>"><img src="img/delete.png" alt="icon delete" width="32" height="32"></a></td>
+                            <td><a href="?delete=<?=$item['idourdatas']?>" onclick="return confirm('Êtes-vous sûre de vouloir le supprimer ?')"><img src="img/delete.png" alt="icon delete" width="32" height="32"></a></td>
                         <?php endforeach; ?>
                     </tr>
                 </table>
